feat: allow match() function on any mock request
