+++ addon/src/mocks/mock-store-request.js
@@ -2,6 +2,7 @@
 /* eslint-disable ember/no-get, ember/classic-decorator-no-classic-methods */
 import FactoryGuy from '../factory-guy';
 import MockRequest from './mock-request';
+import { isEmptyObject, isEquivalent } from '../utils/helper-functions';
 
 export default class extends MockRequest {
   constructor(modelName, requestType) {
@@ -53,5 +54,47 @@ export default class extends MockRequest {
     }
 
     return { adapterOptions: this.adapterOptions, record };
+  }
+
+  /**
+   This is tricky, but the main idea here is:
+
+   #1 Take the keys they want to match and transform them to what the serialized
+   version would be ie. company => company_id
+   #2 Take the matchArgs and turn them into a FactoryGuy payload class by
+   FactoryGuy.build(ing) them into a payload
+   #3 Wrap the request data into a FactoryGuy payload class
+   #4 Go though the keys from #1 and check that both the payloads from #2/#3 have the
+   same values
+
+   @param requestData
+   @returns {boolean} true is no attributes to match or they all match
+   */
+  attributesMatch(requestData, matchParams) {
+    if (isEmptyObject(matchParams)) {
+      return true;
+    }
+
+    let builder = FactoryGuy.fixtureBuilder(this.modelName);
+
+    // transform they match keys
+    let matchCheckKeys = Object.keys(matchParams).map((key) => {
+      return builder.transformKey(this.modelName, key);
+    });
+    // build the match args into a JSONPayload class
+    let buildOpts = { serializeMode: true, transformKeys: true };
+    let expectedData = builder.convertForBuild(
+      this.modelName,
+      matchParams,
+      buildOpts,
+    );
+
+    // wrap request data in a JSONPayload class
+    builder.wrapPayload(this.modelName, requestData);
+
+    // success if all values match
+    return matchCheckKeys.every((key) => {
+      return isEquivalent(expectedData.get(key), requestData.get(key));
+    });
   }
 }
